Move all methods using oc command in cluster package
